pyproject.toml
[tool.poetry]
name = "synthbean-python"
version = "0.1.0"
description = "Elastic Python SynthBean"
authors = ["Mike Place <mike.place@elastic.co>"]

[tool.poetry.dependencies]
python = "^3.9"
elastic-apm = "^6.0.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
